packaging: tame the packaging madness
Cynara seems to adapt the model of multiple small libraries put in separate
package each. This leads into plethora of small packages.
This commit tries to rationalize this decomposition, merging some of the
packages:
- all development packages are now merged into one cynara-devel
- client, client-async and client-common are merged into one
- storage is merged into libcynara-common

The overall result is decrease in total number of packages from 24 to 12,
not counting auto generated debuginfo and debugsource.

Change-Id: I947e733872a4d5c96c722b207243e3c1cdfe1ba6