Drop "method" parameter from creds helper functions
Existing user and client helper functions are over complicated and
difficult to use. The most common usage, and probably the only worth
supporting, would be to rely on global configuration for helper methods.
This use case would require four API calls (and error checking after each)
e.g.:
- cynara_creds_get_default_client_method
- cynara_creds_socket_get_client
- cynara_creds_get_default_user_method
- cynara_creds_socket_get_user

It is also doubtful that the client (a service) should decide which method
to use. Such decision should rather be made system-wide, in the helper
configuration.

This patch removes the method argument and makes the helpers always use
the default method. This takes the freedom away from the clients and
reduces the most useful use case of helpers to 2 API calls.

Change-Id: I7fd5f2e4dc7bde6cd5ecf31e9bbbb0e17b5b3d63
Signed-off-by: Rafal Krypa <r.krypa@samsung.com>