Introduce ChecksumStream
ChecksumStream will replace std::ofstream for saving database contents
in storage. This way data will be not only written to the files, but
also its checksums will be computed and stored in given stream (database
index equivalent for storing checksums).

Checksum computing is performed during stream destruction in order to be
sure that all necessary data was already collected.

Change-Id: I4a9ff2e29361f337cacd790d77364feca854a706