Secure logs format string
This patch makes compilers happy. Some of them tries to shout
if -werror=format-security flag was enabled about unsecure format
string that is created using stringstream.

This patch defines own format string "%s" if there are no additional
parameters except format string passed to LOG... macros.

Dark side of this patch is that usage of %m is not supported
if there are no additional arguments to LOG (beside format message)

Change-Id: I2d4ab5a07d170c85f162a5f59cb63ecbfb56fec6