Fix debug build of chsgen
Debug build break was caused by usage of Cynara Exceptions
that use LOG macros, that in debug mode use Backtrace information.
We cannot link chsgen (a part of migration tool)
with libcynara-commons that provides implementation for Backtrace
because one of migration tool design concept is that it cannot rely
on cynara's libraries, because it is used in time, when cynara packages
are installed, upgraded, downgraded, removed.

That is instead of CynaraExceptions std::runtime_error is used.

Change-Id: Id5447ca769f0bbef19a2f1f216bc672653e736be