Fix deprecation of 'register' keyword in C++11
"register" keyword is deprecated in C++11.
Glib library uses this keyword in its headers.
Including these headers causes a warning to be generated,
which in case of compilation with -Werror flag is elevated to error.

Problem occures only on some compilers e.g. clang 5.1
Problem is fixed by redefinition of "register" keyword
to an empty token in case of compilation for C++11.

Change-Id: I3b8036195ff896ad7a4cd4f42ca5a115401c06ba
Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>