Fix problem with /var/cynara folder creation
Folder is required for locking database to make sure that only daemon or offline api
 acquires access to database.
The folder was formerly created by service rpm, which was required by all other rpm packages.
After releasing version 13.0 the dependency was dropped and folder was not created before usage
 of libraries or cyad.

In this fix folder creation was moved to libcynara-common rpm, which is required by all other rpms.

Change-Id: I7d8513304ec4b84908c4310b23f9a9ff2828d382
Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>