Don't use SMACK label when default method is missing
All credential helpers libraries should follow same convention.
The convention used so far in case of using default client/user method
is to return error, when default configuration file is missing or unable to read.

It is important to notice also, that SMACK might not always be available
on distribution and should not be returned as the default method.

Change-Id: I92be1694b6b78af12c0f17922d7ac62cc82f079c
Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>